requirements.txt
fastapi
mcp
langchain
langchain-mcp-adapters
langchain-openai
langgraph
langfuse
python-dotenv